Moved cryptospecials test into seperate files. rsagen_test.go still has trouble when PEM file for loading does not exist; fails ungracefully. Changed return values to (nil, error) from (privKey, err) in rsagen.go privkeyload.
